// src/app/models/verse.enums.ts
export enum VerseType {
  Hexameter = 'hexameter',
  Pentameter = 'pentameter',
  DistichonHexameter = '+disztichon (hexameter)',
  DistichonPentameter = '+disztichon (pentameter)',
  Unknown = 'unknown form',
}

export enum MetricalDirection {
  Rising = 'emelkedő',
  Falling = 'ereszkedő',
  Mixed = 'vegyes',
}